Create Pattern1
We are currently engaged in an advanced study of Object-Oriented Programming in Java, a distinguished course offered by the esteemed Manipal University Jaipur. Under the expert guidance of Dr. Sourabh Singh, we are delving into the intricacies of this powerful programming paradigm, which promises to enhance our technical acumen and problem-solving skills.